v0.4

+++ sip_assembly/clients.py
@@ -1,5 +1,4 @@
 import base64
-import json
 from os.path import join
 import requests
 
@@ -8,6 +7,7 @@ class ArchivematicaClientException(Exception): pass
 
 
 class ArchivematicaClient(object):
+    """Client for interacting with Archivematica."""
     def __init__(self, username, api_key, baseurl, location_uuid):
         self.headers = {"Authorization": "ApiKey {}:{}".format(username, api_key)}
         self.baseurl = baseurl
@@ -16,12 +16,15 @@ class ArchivematicaClient(object):
     def retrieve(self, uri, *args, **kwargs):
         full_url = "/".join([self.baseurl.rstrip("/"), uri.lstrip("/")])
         response = requests.get(full_url, headers=self.headers, *args, **kwargs)
-        if response:
-            return response
+        if response.status_code == 200:
+            if response.headers['content-type'] == 'application/json':
+                return response.json()
+            else:
+                return response
         else:
             raise ArchivematicaClientException("Could not return a valid response for {}".format(full_url))
 
-    def send_start_transfer_request(self, sip):
+    def start_transfer(self, sip):
         """Starts a transfer in Archivematica."""
         basepath = "{}.tar.gz".format(sip.bag_identifier)
         full_url = join(self.baseurl, 'transfer/start_transfer/')
@@ -33,7 +36,7 @@ class ArchivematicaClient(object):
             message = start.json()['message'] if start.json()['message'] else start.reason
             raise ArchivematicaClientException(message)
 
-    def send_approve_transfer_request(self, sip):
+    def approve_transfer(self, sip):
         """Approves a transfer in Archivematica."""
         approve_transfer = requests.post(join(self.baseurl, 'transfer/approve_transfer/'),
                                          headers=self.headers,
@@ -41,22 +44,14 @@ class ArchivematicaClient(object):
         if approve_transfer.status_code != 200:
             raise ArchivematicaClientException(approve_transfer.json()['message'])
 
-    def send_ingest_cleanup_request(self):
-        """Removes completed ingests."""
-        completed = self.retrieve('ingest/completed').json()
+    def cleanup(self, type):
+        """Removes completed ingests and transfers in Archivematica."""
+        if type not in ["ingest", "transfer"]:
+            raise ArchivematicaClientException("Unknown type {}".format(type))
+        completed = self.retrieve('{}/completed'.format(type))
         for uuid in completed['results']:
-            full_url = join(self.baseurl, 'ingest/{}/delete/'.format(uuid))
+            full_url = join(self.baseurl, '{}/{}/delete/'.format(type, uuid))
             resp = requests.delete(full_url, headers=self.headers).json()
             if not resp['removed']:
-                raise ArchivematicaClientException("Error removing ingest {}".format(uuid))
-        return len(completed['results'])
-
-    def send_transfer_cleanup_request(self):
-        """Removes completed transfers."""
-        completed = self.retrieve('transfer/completed').json()
-        for uuid in completed['results']:
-            full_url = join(self.baseurl, 'transfer/{}/delete/'.format(uuid))
-            resp = requests.delete(full_url, headers=self.headers).json()
-            if not resp['removed']:
-                raise ArchivematicaClientException("Error removing transfer {}".format(uuid))
-        return len(completed['results'])
+                raise ArchivematicaClientException(uuid)
+        return completed['results']
